remove detector position if present and make loki at larmor submodule
