+++ src/ess/loki/workflow.py
@@ -10,40 +10,24 @@ import scippnexus as snx
 
 from ess import sans
 from ess.reduce.workflow import register_workflow
-from ess.sans.io import read_xml_detector_masking
 from ess.sans.parameters import typical_outputs
 
-from ..sans.conversions import ElasticCoordTransformGraph
 from ..sans.types import (
-    BackgroundRun,
-    BeamCenter,
     DetectorBankSizes,
     DetectorPixelShape,
     DirectBeam,
     DirectBeamFilename,
-    EmptyBeamRun,
-    Filename,
     Incident,
     LookupTableRelativeErrorThreshold,
-    MonitorType,
     NeXusComponent,
-    NeXusDetectorName,
     NeXusMonitorName,
     NonBackgroundWavelengthRange,
-    PixelMaskFilename,
     PixelShapePath,
-    RawDetector,
-    RawMonitor,
     RunType,
-    SampleRun,
     Transmission,
-    TransmissionRun,
-    WavelengthDetector,
-    WavelengthMonitor,
 )
 
 DETECTOR_BANK_SIZES = {
-    'larmor_detector': {'layer': 4, 'tube': -1, 'straw': 7, 'pixel': 512},
     'loki_detector_0': {'layer': 4, 'tube': 56, 'straw': 7, 'pixel': -1},
     'loki_detector_1': {'layer': 4, 'tube': 16, 'straw': 7, 'pixel': -1},
     'loki_detector_2': {'layer': 4, 'tube': 12, 'straw': 7, 'pixel': -1},
@@ -54,16 +38,6 @@ DETECTOR_BANK_SIZES = {
     'loki_detector_7': {'layer': 4, 'tube': 20, 'straw': 7, 'pixel': -1},
     'loki_detector_8': {'layer': 4, 'tube': 32, 'straw': 7, 'pixel': -1},
 }
-
-
-def larmor_default_parameters() -> dict:
-    return {
-        DetectorBankSizes: DETECTOR_BANK_SIZES,
-        NeXusMonitorName[Incident]: 'monitor_1',
-        NeXusMonitorName[Transmission]: 'monitor_2',
-        PixelShapePath: 'pixel_shape',
-        NonBackgroundWavelengthRange: None,
-    }
 
 
 def loki_default_parameters() -> dict:
@@ -83,45 +57,6 @@ def loki_default_parameters() -> dict:
     }
 
 
-def _larmor_convert_to_wavelength(da: sc.DataArray, graph: dict) -> sc.DataArray:
-    event_time_offset = da.bins.coords['event_time_offset']
-    da = da.bins.drop_coords('event_time_offset')
-    da.bins.coords['tof'] = event_time_offset
-    if 'event_time_zero' in da.dims:
-        da = da.bins.concat('event_time_zero')
-    return da.transform_coords('wavelength', graph=graph, keep_intermediate=False)
-
-
-def larmor_data_to_wavelength(
-    da: RawDetector[RunType], graph: ElasticCoordTransformGraph[RunType]
-) -> WavelengthDetector[RunType]:
-    """
-    Compute wavelength coordinate for Loki detector data at Larmor.
-    This is different from the standard conversion from the GenericUnwrapWorkflow
-    because the detector test was conducted as ISIS where the pulse has a different time
-    structure.
-    The conversion here is much simpler: the event_time_offset coordinate is directly
-    renamed as time-of-flight, and a wavelength is computed from that.
-    """
-    return WavelengthDetector[RunType](_larmor_convert_to_wavelength(da, graph=graph))
-
-
-def larmor_monitor_to_wavelength(
-    da: RawMonitor[RunType, MonitorType], graph: ElasticCoordTransformGraph[RunType]
-) -> WavelengthMonitor[RunType, MonitorType]:
-    """
-    Compute wavelength coordinate for Loki monitor data at Larmor.
-    This is different from the standard conversion from the GenericUnwrapWorkflow
-    because the detector test was conducted as ISIS where the pulse has a different time
-    structure.
-    The conversion here is much simpler: the event_time_offset coordinate is directly
-    renamed as time-of-flight, and a wavelength is computed from that.
-    """
-    return WavelengthMonitor[RunType, MonitorType](
-        _larmor_convert_to_wavelength(da, graph=graph)
-    )
-
-
 def detector_pixel_shape(
     detector: NeXusComponent[snx.NXdetector, RunType],
     pixel_shape_path: PixelShapePath,
@@ -135,54 +70,6 @@ def load_direct_beam(filename: DirectBeamFilename) -> DirectBeam:
 
 
 loki_providers = (detector_pixel_shape, load_direct_beam)
-
-
-@register_workflow
-def LokiAtLarmorWorkflow() -> sciline.Pipeline:
-    """
-    Workflow with default parameters for Loki test at Larmor.
-
-    This version of the Loki workflow:
-
-    - Uses ISIS XML files to define masks.
-    - Sets a dummy sample position [0,0,0] since files do not contain this information.
-
-    Returns
-    -------
-    :
-        Loki workflow as a sciline.Pipeline
-    """
-    workflow = sans.SansWorkflow()
-    for provider in loki_providers:
-        workflow.insert(provider)
-    for key, param in larmor_default_parameters().items():
-        workflow[key] = param
-    workflow.insert(larmor_data_to_wavelength)
-    workflow.insert(larmor_monitor_to_wavelength)
-    workflow.insert(read_xml_detector_masking)
-    workflow[NeXusDetectorName] = 'larmor_detector'
-    workflow.typical_outputs = typical_outputs
-    return workflow
-
-
-@register_workflow
-def LokiAtLarmorTutorialWorkflow() -> sciline.Pipeline:
-    from ess.loki import data
-
-    workflow = LokiAtLarmorWorkflow()
-
-    workflow[PixelMaskFilename] = list(map(str, data.loki_tutorial_mask_filenames()))
-    workflow[Filename[SampleRun]] = str(data.loki_tutorial_sample_run_60339())
-    workflow[Filename[BackgroundRun]] = str(data.loki_tutorial_background_run_60393())
-    workflow[Filename[TransmissionRun[SampleRun]]] = str(
-        data.loki_tutorial_sample_transmission_run()
-    )
-    workflow[Filename[TransmissionRun[BackgroundRun]]] = str(
-        data.loki_tutorial_run_60392()
-    )
-    workflow[Filename[EmptyBeamRun]] = str(data.loki_tutorial_run_60392())
-    workflow[BeamCenter] = sc.vector(value=[-0.02914868, -0.01816138, 0.0], unit='m')
-    return workflow
 
 
 @register_workflow
